Correction made for writing date values and number values.
When the datetime has the Unspecified Kind, the string didn't contain any timezone. Changed this to local time if unspecified for DateField and RelevantDate on Write.

When the NumberField had a currencyCode, do not pass the numberStyle. It's one or the other. When currencyCode is passed, the numberStyle is ignored on Write.